move parameters() to DeleteReadTests
