+++ data/src/test/java/org/apache/iceberg/data/DeleteReadTests.java
@@ -28,7 +28,10 @@ import java.util.List;
 import java.util.Set;
 import org.apache.iceberg.DataFile;
 import org.apache.iceberg.DeleteFile;
+import org.apache.iceberg.FileFormat;
 import org.apache.iceberg.Files;
+import org.apache.iceberg.Parameter;
+import org.apache.iceberg.Parameters;
 import org.apache.iceberg.PartitionSpec;
 import org.apache.iceberg.Schema;
 import org.apache.iceberg.Table;
@@ -76,6 +79,17 @@ public abstract class DeleteReadTests {
   protected List<Record> records = null;
   private List<Record> dateRecords = null;
   protected DataFile dataFile = null;
+
+  @Parameter protected FileFormat format;
+
+  @Parameters(name = "fileFormat = {0}")
+  public static Object[][] parameters() {
+    return new Object[][] {
+      new Object[] {FileFormat.PARQUET},
+      new Object[] {FileFormat.AVRO},
+      new Object[] {FileFormat.ORC}
+    };
+  }
 
   @BeforeEach
   public void writeTestDataFile() throws IOException {
